# src/main/java/com/aslan/project/bank_rest/dto/request/RegisterRequest.java
package com.aslan.project.bank_rest.dto.request;

import jakarta.validation.constraints.NotBlank;
import lombok.Getter;
import lombok.Setter;

@Getter @Setter
public class RegisterRequest {
    @NotBlank private String username;
    @NotBlank private String password;
    @NotBlank private String role;

}